Change Log Update (00.00.47)
* 2016-09-17
+ Worked more on the sign-up process,  Now saving user record and
security record
to the Sqlite database.  Still need a little work on the dates and best
way
to store and retrieve UTC date and times.
+ Also cleaned up the mod_signup code some and optimized the
displayPrompts

+++ src/model/users.hpp
@@ -93,7 +93,7 @@ public:
     bool bIgnoreTimeLimit;
     bool bAllowPurge;
 
-    int iSecurityIndex;
+    long iSecurityIndex;
 
     // Default User ID to -1 so we know it's not active!
     Users()
